+++ lib/contentful/bootstrap/templates.rb
@@ -1,0 +1,3 @@
+require "contentful/bootstrap/templates/blog"
+require "contentful/bootstrap/templates/catalogue"
+require "contentful/bootstrap/templates/gallery"