+++ lib/contentful/bootstrap/templates/links/asset.rb
@@ -1,0 +1,15 @@
+require "contentful/bootstrap/templates/links/base"
+
+module Contentful
+  module Bootstrap
+    module Templates
+      module Links
+        class Asset < Base
+          def kind
+            :assets
+          end
+        end
+      end
+    end
+  end
+end