+++ lib/contentful_bootstrap/constants.rb
@@ -1,6 +1,0 @@
-module ContentfulBootstrap
-  module Constants
-    OAUTH_APP_ID = "a19770bea126e6d596d8599ff42e14173409e3e252895b78d0cb289c10586276".freeze
-    OAUTH_CALLBACK_URL = "http://localhost:5123/oauth_callback".freeze
-  end
-end